NAS-115487 / 22.12 / SMB form refactor (#6605)

+++ src/app/services/dialog.service.ts
@@ -67,7 +67,7 @@ export class DialogService {
       dialogRef.componentInstance.method = options.method;
       dialogRef.componentInstance.switchSelectionEmitter.pipe(untilDestroyed(this)).subscribe((selection: boolean) => {
         const data = options.data;
-        if (selection && data[0]) {
+        if (selection && data && data[0]) {
           if (data[0] && data[0].hasOwnProperty('reboot')) {
             data[0].reboot = !data[0].reboot;
           }
